Move exceptions to the package root
