+++ hammurabi/exceptions.py
